python-app/requirements.txt
Flask==1.1.1
pymongo==3.10.1
Flask_Caching==1.8.0
Flask_Cors==3.0.8
request==2019.4.13
kafka-python
